refactor: Rename lend to earn
