+++ apps/web/src/config/routes.ts
@@ -8,10 +8,10 @@ export const AppRoutes = {
   stake: '/stake',
   privacy: '/privacy',
   licenses: '/licenses',
-  lend: '/lend',
   index: '/',
   imprint: '/imprint',
   home: '/home',
+  earn: '/earn',
   cookie: '/cookie',
   bridge: '/bridge',
   addressBook: '/address-book',
