+++ apps/web/src/features/earn/components/VaultDeposit/index.tsx
