+++ apps/web/src/features/earn/constants.ts
@@ -1,10 +1,10 @@
-export const LEND_TITLE = 'Lend'
+export const EARN_TITLE = 'Lend'
 export const WIDGET_TESTNET_URL = 'https://safe.widget.testnet.kiln.fi/overview'
-export const LEND_CONSENT_STORAGE_KEY = 'lendDisclaimerAcceptedV1'
+export const EARN_CONSENT_STORAGE_KEY = 'lendDisclaimerAcceptedV1'
 
 export const widgetAppData = {
   url: WIDGET_TESTNET_URL,
-  name: LEND_TITLE,
+  name: EARN_TITLE,
   iconUrl: '/images/common/stake.svg',
   chainIds: ['17000', '11155111', '1', '42161', '137', '56', '8453', '10'],
 }
