+++ apps/web/src/features/earn/hooks/useGetWidgetUrl.ts
