+++ apps/web/src/features/earn/index.tsx
@@ -4,12 +4,12 @@ import Disclaimer from '@/components/common/Disclaimer'
 import WidgetDisclaimer from '@/components/common/WidgetDisclaimer'
 import BlockedAddress from '@/components/common/BlockedAddress'
 import useBlockedAddress from '@/hooks/useBlockedAddress'
-import LendWidget from '@/features/lend/components/LendWidget'
+import EarnWidget from 'src/features/earn/components/EarnWidget'
 import useConsent from '@/hooks/useConsent'
-import { LEND_CONSENT_STORAGE_KEY } from '@/features/lend/constants'
+import { EARN_CONSENT_STORAGE_KEY } from '@/features/earn/constants'
 
-const LendPage = () => {
-  const { isConsentAccepted, onAccept } = useConsent(LEND_CONSENT_STORAGE_KEY)
+const EarnPage = () => {
+  const { isConsentAccepted, onAccept } = useConsent(EARN_CONSENT_STORAGE_KEY)
   const router = useRouter()
   const { asset } = router.query
 
@@ -35,7 +35,7 @@ const LendPage = () => {
   return (
     <>
       {isConsentAccepted ? (
-        <LendWidget asset={String(asset)} />
+        <EarnWidget asset={String(asset)} />
       ) : (
         <Stack
           direction="column"
@@ -57,4 +57,4 @@ const LendPage = () => {
   )
 }
 
-export default LendPage
+export default EarnPage
